@@ -2,9 +2,9 @@
 /* eslint-disable */
 /**
  * NLP Sandbox PHI Deidentifier API
- * # Introduction This NLP tool takes as input a clinical note and returned a de-identified version of the note. This design of this API is a work in progress. # Examples - [NLP Sandbox PHI Deidentifier](https://github.com/nlpsandbox/phi-deidentifier) 
+ * # Introduction This NLP tool takes as input a clinical note and returned a de-identified version of the note. This design of this API is a work in progress. # Examples - [NLP Sandbox PHI Deidentifier Example](https://github.com/nlpsandbox/phi-deidentifier-example) 
  *
- * The version of the OpenAPI document: 1.1.2
+ * The version of the OpenAPI document: 1.2.0
  * Contact: team@nlpsandbox.io
  *
  * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
@@ -26,14 +26,14 @@ import {
     TextIdAnnotationFromJSON,
     TextIdAnnotationFromJSONTyped,
     TextIdAnnotationToJSON,
+    TextLocationAnnotation,
+    TextLocationAnnotationFromJSON,
+    TextLocationAnnotationFromJSONTyped,
+    TextLocationAnnotationToJSON,
     TextPersonNameAnnotation,
     TextPersonNameAnnotationFromJSON,
     TextPersonNameAnnotationFromJSONTyped,
     TextPersonNameAnnotationToJSON,
-    TextPhysicalAddressAnnotation,
-    TextPhysicalAddressAnnotationFromJSON,
-    TextPhysicalAddressAnnotationFromJSONTyped,
-    TextPhysicalAddressAnnotationToJSON,
 } from './';
 
 /**
@@ -55,11 +55,11 @@ export interface AnnotationSet {
      */
     textPersonNameAnnotations: Array<TextPersonNameAnnotation>;
     /**
-     * Physical address annotations in a text
-     * @type {Array<TextPhysicalAddressAnnotation>}
+     * Location annotations in a text
+     * @type {Array<TextLocationAnnotation>}
      * @memberof AnnotationSet
      */
-    textPhysicalAddressAnnotations: Array<TextPhysicalAddressAnnotation>;
+    textLocationAnnotations: Array<TextLocationAnnotation>;
     /**
      * ID annotations in a text
      * @type {Array<TextIdAnnotation>}
@@ -86,7 +86,7 @@ export function AnnotationSetFromJSONTyped(json: any, ignoreDiscriminator: boole
         
         'textDateAnnotations': ((json['textDateAnnotations'] as Array<any>).map(TextDateAnnotationFromJSON)),
         'textPersonNameAnnotations': ((json['textPersonNameAnnotations'] as Array<any>).map(TextPersonNameAnnotationFromJSON)),
-        'textPhysicalAddressAnnotations': ((json['textPhysicalAddressAnnotations'] as Array<any>).map(TextPhysicalAddressAnnotationFromJSON)),
+        'textLocationAnnotations': ((json['textLocationAnnotations'] as Array<any>).map(TextLocationAnnotationFromJSON)),
         'textIdAnnotations': ((json['textIdAnnotations'] as Array<any>).map(TextIdAnnotationFromJSON)),
         'textContactAnnotations': ((json['textContactAnnotations'] as Array<any>).map(TextContactAnnotationFromJSON)),
     };
@@ -103,7 +103,7 @@ export function AnnotationSetToJSON(value?: AnnotationSet | null): any {
         
         'textDateAnnotations': ((value.textDateAnnotations as Array<any>).map(TextDateAnnotationToJSON)),
         'textPersonNameAnnotations': ((value.textPersonNameAnnotations as Array<any>).map(TextPersonNameAnnotationToJSON)),
-        'textPhysicalAddressAnnotations': ((value.textPhysicalAddressAnnotations as Array<any>).map(TextPhysicalAddressAnnotationToJSON)),
+        'textLocationAnnotations': ((value.textLocationAnnotations as Array<any>).map(TextLocationAnnotationToJSON)),
         'textIdAnnotations': ((value.textIdAnnotations as Array<any>).map(TextIdAnnotationToJSON)),
         'textContactAnnotations': ((value.textContactAnnotations as Array<any>).map(TextContactAnnotationToJSON)),
     };
