@@ -2,9 +2,9 @@
 /* eslint-disable */
 /**
  * NLP Sandbox PHI Deidentifier API
- * # Introduction This NLP tool takes as input a clinical note and returned a de-identified version of the note. This design of this API is a work in progress. # Examples - [NLP Sandbox PHI Deidentifier](https://github.com/nlpsandbox/phi-deidentifier) 
+ * # Introduction This NLP tool takes as input a clinical note and returned a de-identified version of the note. This design of this API is a work in progress. # Examples - [NLP Sandbox PHI Deidentifier Example](https://github.com/nlpsandbox/phi-deidentifier-example) 
  *
- * The version of the OpenAPI document: 1.1.2
+ * The version of the OpenAPI document: 1.2.0
  * Contact: team@nlpsandbox.io
  *
  * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
@@ -43,7 +43,7 @@ export interface DeidentificationStep {
      */
     maskingCharConfig?: MaskingCharConfig;
     /**
-     * Configuration for the "annotation type" strategy. E.g. "John Smith lives at 123 Main St" -> "[PERSON_NAME] lives at [PHYSICAL_ADDRESS]".
+     * Configuration for the "annotation type" strategy. E.g. "John Smith lives at 123 Main St" -> "[PERSON_NAME] lives at [LOCATION]".
      * @type {object}
      * @memberof DeidentificationStep
      */
@@ -75,7 +75,7 @@ export interface DeidentificationStep {
 export enum DeidentificationStepAnnotationTypesEnum {
     Date = 'text_date',
     PersonName = 'text_person_name',
-    PhysicalAddress = 'text_physical_address',
+    Location = 'text_location',
     Id = 'text_id',
     Contact = 'text_contact'
 }
