@@ -2,9 +2,9 @@
 /* eslint-disable */
 /**
  * NLP Sandbox PHI Deidentifier API
- * # Introduction This NLP tool takes as input a clinical note and returned a de-identified version of the note. This design of this API is a work in progress. # Examples - [NLP Sandbox PHI Deidentifier](https://github.com/nlpsandbox/phi-deidentifier) 
+ * # Introduction This NLP tool takes as input a clinical note and returned a de-identified version of the note. This design of this API is a work in progress. # Examples - [NLP Sandbox PHI Deidentifier Example](https://github.com/nlpsandbox/phi-deidentifier-example) 
  *
- * The version of the OpenAPI document: 1.1.2
+ * The version of the OpenAPI document: 1.2.0
  * Contact: team@nlpsandbox.io
  *
  * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
@@ -18,55 +18,55 @@ import {
     TextAnnotationFromJSON,
     TextAnnotationFromJSONTyped,
     TextAnnotationToJSON,
-    TextPhysicalAddressAnnotationAllOf,
-    TextPhysicalAddressAnnotationAllOfFromJSON,
-    TextPhysicalAddressAnnotationAllOfFromJSONTyped,
-    TextPhysicalAddressAnnotationAllOfToJSON,
+    TextLocationAnnotationAllOf,
+    TextLocationAnnotationAllOfFromJSON,
+    TextLocationAnnotationAllOfFromJSONTyped,
+    TextLocationAnnotationAllOfToJSON,
 } from './';
 
 /**
- * A physical address annotation in a text
+ * A location annotation in a text
  * @export
- * @interface TextPhysicalAddressAnnotation
+ * @interface TextLocationAnnotation
  */
-export interface TextPhysicalAddressAnnotation {
+export interface TextLocationAnnotation {
     /**
      * The position of the first character
      * @type {number}
-     * @memberof TextPhysicalAddressAnnotation
+     * @memberof TextLocationAnnotation
      */
     start: number;
     /**
      * The length of the annotation
      * @type {number}
-     * @memberof TextPhysicalAddressAnnotation
+     * @memberof TextLocationAnnotation
      */
     length: number;
     /**
      * The string annotated
      * @type {string}
-     * @memberof TextPhysicalAddressAnnotation
+     * @memberof TextLocationAnnotation
      */
     text: string;
     /**
      * The confidence in the accuracy of the annotation
      * @type {number}
-     * @memberof TextPhysicalAddressAnnotation
+     * @memberof TextLocationAnnotation
      */
     confidence: number;
     /**
-     * Type of address information
+     * Type of location
      * @type {string}
-     * @memberof TextPhysicalAddressAnnotation
+     * @memberof TextLocationAnnotation
      */
-    addressType: TextPhysicalAddressAnnotationAddressTypeEnum;
+    locationType: TextLocationAnnotationLocationTypeEnum;
 }
 
 /**
 * @export
 * @enum {string}
 */
-export enum TextPhysicalAddressAnnotationAddressTypeEnum {
+export enum TextLocationAnnotationLocationTypeEnum {
     City = 'city',
     Country = 'country',
     Department = 'department',
@@ -79,11 +79,11 @@ export enum TextPhysicalAddressAnnotationAddressTypeEnum {
     Zip = 'zip'
 }
 
-export function TextPhysicalAddressAnnotationFromJSON(json: any): TextPhysicalAddressAnnotation {
-    return TextPhysicalAddressAnnotationFromJSONTyped(json, false);
+export function TextLocationAnnotationFromJSON(json: any): TextLocationAnnotation {
+    return TextLocationAnnotationFromJSONTyped(json, false);
 }
 
-export function TextPhysicalAddressAnnotationFromJSONTyped(json: any, ignoreDiscriminator: boolean): TextPhysicalAddressAnnotation {
+export function TextLocationAnnotationFromJSONTyped(json: any, ignoreDiscriminator: boolean): TextLocationAnnotation {
     if ((json === undefined) || (json === null)) {
         return json;
     }
@@ -93,11 +93,11 @@ export function TextPhysicalAddressAnnotationFromJSONTyped(json: any, ignoreDisc
         'length': json['length'],
         'text': json['text'],
         'confidence': json['confidence'],
-        'addressType': json['addressType'],
+        'locationType': json['locationType'],
     };
 }
 
-export function TextPhysicalAddressAnnotationToJSON(value?: TextPhysicalAddressAnnotation | null): any {
+export function TextLocationAnnotationToJSON(value?: TextLocationAnnotation | null): any {
     if (value === undefined) {
         return undefined;
     }
@@ -110,7 +110,7 @@ export function TextPhysicalAddressAnnotationToJSON(value?: TextPhysicalAddressA
         'length': value.length,
         'text': value.text,
         'confidence': value.confidence,
-        'addressType': value.addressType,
+        'locationType': value.locationType,
     };
 }
 
