@@ -2,9 +2,9 @@
 /* eslint-disable */
 /**
  * NLP Sandbox PHI Deidentifier API
- * # Introduction This NLP tool takes as input a clinical note and returned a de-identified version of the note. This design of this API is a work in progress. # Examples - [NLP Sandbox PHI Deidentifier](https://github.com/nlpsandbox/phi-deidentifier) 
+ * # Introduction This NLP tool takes as input a clinical note and returned a de-identified version of the note. This design of this API is a work in progress. # Examples - [NLP Sandbox PHI Deidentifier Example](https://github.com/nlpsandbox/phi-deidentifier-example) 
  *
- * The version of the OpenAPI document: 1.1.2
+ * The version of the OpenAPI document: 1.2.0
  * Contact: team@nlpsandbox.io
  *
  * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
@@ -18,6 +18,10 @@ import {
     LicenseFromJSON,
     LicenseFromJSONTyped,
     LicenseToJSON,
+    ToolType,
+    ToolTypeFromJSON,
+    ToolTypeFromJSONTyped,
+    ToolTypeToJSON,
 } from './';
 
 /**
@@ -75,11 +79,11 @@ export interface Tool {
      */
     url: string;
     /**
-     * The type of this tool
-     * @type {string}
+     * 
+     * @type {ToolType}
      * @memberof Tool
      */
-    type: string;
+    type: ToolType;
     /**
      * The version of the tool OpenAPI specification
      * @type {string}
@@ -106,7 +110,7 @@ export function ToolFromJSONTyped(json: any, ignoreDiscriminator: boolean): Tool
         'author': json['author'],
         'authorEmail': json['authorEmail'],
         'url': json['url'],
-        'type': json['type'],
+        'type': ToolTypeFromJSON(json['type']),
         'apiVersion': json['apiVersion'],
     };
 }
@@ -128,7 +132,7 @@ export function ToolToJSON(value?: Tool | null): any {
         'author': value.author,
         'authorEmail': value.authorEmail,
         'url': value.url,
-        'type': value.type,
+        'type': ToolTypeToJSON(value.type),
         'apiVersion': value.apiVersion,
     };
 }
