@@ -15,8 +15,9 @@ export * from './TextDateAnnotation';
 export * from './TextDateAnnotationAllOf';
 export * from './TextIdAnnotation';
 export * from './TextIdAnnotationAllOf';
+export * from './TextLocationAnnotation';
+export * from './TextLocationAnnotationAllOf';
 export * from './TextPersonNameAnnotation';
-export * from './TextPhysicalAddressAnnotation';
-export * from './TextPhysicalAddressAnnotationAllOf';
 export * from './Tool';
 export * from './ToolDependencies';
+export * from './ToolType';
